✨ Refactor: Extract timer component and utility functions for focus page
